Fix initial problems with Viterbi algo

+++ semi_markov/accelerator.py
@@ -18,9 +18,9 @@ def viterbi(trans_mat: multidim,
             ) -> singldim:
 
     buff_p_c = c.POINTER(c.c_int)
-    out = np.array(out_state_seq).astype(int)
+    out = np.array(out_state_seq).astype(np.int32)
     buff_p = out.ctypes.data_as(buff_p_c)
-    
+
     lib.viterbi(c_array(np.array(trans_mat,dtype=float)),
                 c_array(np.array(init_state,dtype=float)), c.c_size_t(state_num),
                 c_array(np.array(obs_probs,dtype=float)), c.c_size_t(time_len),
